requirements.txt
blinker==1.6.3
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==21.2.0
importlib-metadata ==6.8.0
itsdangerous==2.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
python-dotenv==1.0.0
werkzeug==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
zipp==3.17.0
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
SQLAlchemy==2.0.31